Extract web_poet.serialization.utils.
